Check to see if conf file already exists before installing it. (#191)
* Update install.sh

* Add existing conf path.

+++ install.sh
@@ -9,7 +9,9 @@ install -m 644 log2ram.service /etc/systemd/system/log2ram.service
 install -m 644 log2ram-daily.service /etc/systemd/system/log2ram-daily.service
 install -m 644 log2ram-daily.timer /etc/systemd/system/log2ram-daily.timer
 install -m 755 log2ram /usr/local/bin/log2ram
-install -m 644 log2ram.conf /etc/log2ram.conf
+if [ ! -f /etc/log2ram.conf ]; then
+	install -m 644 log2ram.conf /etc/log2ram.conf
+fi
 install -m 644 uninstall.sh /usr/local/bin/uninstall-log2ram.sh
 systemctl enable log2ram.service log2ram-daily.timer
